Вторая ветка через аннотации.

+++ src/main/java/by/sTm/target1/CurrencyExchange.java
@@ -3,5 +3,5 @@ package by.sTm.target1;
 import java.math.BigDecimal;
 
 public interface CurrencyExchange {
-    String getExchange(BigDecimal byn);
+    String getExchange(int byn);
 }
